backend-api/pyproject.toml
[project]
name = "backend-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "azure-cosmos",
    "pyjwt",
    "passlib[bcrypt]",
    "pydantic ",
    "python-jose",
    "pydantic[email]"
]
